+++ src/commands/hardis/package/install.ts
@@ -10,7 +10,8 @@ import * as path from 'path';
 import * as prompts from 'prompts';
 // import * as packages from '../../../../defaults/packages.json'
 import { MetadataUtils } from '../../../common/metadata-utils';
-import { execCommand, isCI, uxLog } from '../../../common/utils';
+import { isCI, uxLog } from '../../../common/utils';
+import { getConfig, setConfig } from '../../../config';
 
 // Initialize Messages with the current plugin directory
 Messages.importMessagesDirectory(__dirname);
@@ -49,10 +50,12 @@ export default class PackageVersionInstall extends SfdxCommand {
 
     /* jscpd:ignore-end */
 
+    protected allPackagesFileName = path.join(__dirname, './../../../../defaults/packages.json');
+    protected sfdxProjectJsonFileName = path.join(process.cwd(), 'sfdx-project.json');
+
     public async run(): Promise<AnyJson> {
-        const packagesRaw = await fs.readFile(path.join(__dirname, './../../../../defaults/packages.json'), 'utf8');
+        const packagesRaw = await fs.readFile(this.allPackagesFileName, 'utf8');
         const packages = JSON.parse(packagesRaw);
-        const debugMode = this.flags.debug || false;
         const packageId = this.flags.package || null;
         const packagesToInstall = [];
         // If no package Id is sent, ask user what package he/she wants to install
@@ -90,16 +93,48 @@ export default class PackageVersionInstall extends SfdxCommand {
         }));
         // Install packages
         await MetadataUtils.installPackagesOnOrg(packagesToInstallCompleted, null, this);
+        const installedPackages = await MetadataUtils.listInstalledPackages(null, this);
+        uxLog(this, c.italic(c.grey('New package list on org:\n' + JSON.stringify(installedPackages, null, 2))));
 
-        // Retrieve package sources if requested
         if (!isCI) {
-            const pullResponse = await prompts({
-                type: 'confirm',
-                name: 'value',
-                message: c.cyanBright('Do you want to retrieve installed package sources in your local branch ?'),
-                initial: true
-            });
-            if (pullResponse.value === true) {
+            // Add package installation to project .sfdx-hardis.yml
+            const config = await getConfig('project');
+            const projectPackages = config.installedPackages || [];
+            let updated = false;
+            for (const installedPackage of installedPackages) {
+                const matchInstalled = packagesToInstallCompleted.filter(pckg => pckg.key === installedPackage.SubscriberPackageName);
+                const matchLocal = projectPackages.filter(projectPackage => installedPackage.SubscriberPackageName === projectPackage.SubscriberPackageName);
+                if (matchInstalled.length > 0 && matchLocal.length === 0) {
+                    projectPackages.push(installedPackage);
+                    updated = true;
+                }
+            }
+            if (updated) {
+                uxLog(this, 'Updating project sfdx-hardis config to packages are installed everytime');
+                await setConfig('project', { installedPackages: projectPackages });
+            }
+        }
+
+        /* disabled until sfdx multiple package deployment is working >_<
+        // Post install actions
+        if (!isCI && fs.existsSync(this.sfdxProjectJsonFileName)) {
+
+            const postInstallResponse = await prompts([
+                {
+                    type: 'confirm',
+                    name: 'retrieve',
+                    message: c.cyanBright('Do you want to retrieve installed package sources in your local branch ?'),
+                    initial: true
+                },
+                {
+                    type: 'confirm',
+                    name: 'sfdxProject',
+                    message: c.cyanBright('Do you want to update your sfdx-project.json ? (advice: yes)'),
+                    initial: true
+                }
+            ]);
+            // Retrieve package sources if requested
+            if (postInstallResponse.retrieve === true) {
                 for (const pckg of packagesToInstallCompleted) {
                     const retrieveCommand = 'sfdx force:source:retrieve' +
                     ` -n ${pckg.key}` +
@@ -116,7 +151,26 @@ export default class PackageVersionInstall extends SfdxCommand {
                     }
                 }
             }
+            // Update sfdx-project.json with new unlocked packages folder references, so it is taken in account with force:source:push and force:source:pull
+            if (postInstallResponse.sfdxProject === true) {
+                const sfdxProjectRaw = await fs.readFile(this.sfdxProjectJsonFileName, 'utf8');
+                const sfdxProject = JSON.parse(sfdxProjectRaw);
+                let updated = false;
+                for (const installedPackage of installedPackages) {
+                    const matchInstalled = packagesToInstallCompleted.filter(pckg => pckg.key === installedPackage.SubscriberPackageName);
+                    const matchLocal = sfdxProject.packageDirectories.filter(packageDirectory => installedPackage.SubscriberPackageName === packageDirectory.path);
+                    if (matchInstalled.length > 0 && matchLocal.length === 0) {
+                        sfdxProject.packageDirectories.push({path: installedPackage.SubscriberPackageName});
+                        updated = true;
+                    }
+                }
+                if (updated) {
+                    await fs.writeFile(this.sfdxProjectJsonFileName, JSON.stringify(sfdxProject, null, 2));
+                    uxLog(this, c.cyan('[config] Updated sfdx-project.json to add new package folders'));
+                }
+            }
         }
+        */
 
         // Return an object to be displayed with --json
         return { outputString: 'Installed package(s)' };